Move EndpointInfo into iroh-dns as well.
This allows some more things to be come private and avoids an Ext trait.
